Create localization.txt
Hello, big thx for the mods ;). The creature pack is OK. I don't find any probs with it. I crossed on my way the big spider, the spheric robot, the itinerant trader, the santa zombie, the snowfox, and the goat :D they works magnifically and strongs !!.

The file I put is the localization.txt for the mod "junk items". I correct some errors in english and add the translation in French. All is ok for now. I don't find anything wrong. So you can put it in the mod and paste the file actually in the mod. Like this it would write all is in the mod in 2 languages.

I had to my game and test SteelBars and ZombiePinups and they works fine ;).

Keep the good work ;)
Rygastan